chore: move createKeys to the crypto lib
