+++ packages/element-lib/src/func/index.js
@@ -1,7 +1,6 @@
 const func = require('./func');
 
 module.exports = {
-  createKeys: require('./createKeys'),
   objectToMultiHash: require('./objectToMultihash'),
   ...func,
 };
